improve mockup code + rename GetAsQueryable<T> into GetQueryable<T> to match the CoreAPI method naming
